+++ gradle.properties
@@ -1,3 +1,3 @@
-projectName        = GlintlessPotions
+projectName        = ModernPotions
 projectGroup       = me.zircta
 projectVersion     = 2.0
