+++ src/main/resources/weave.mod.json
@@ -1,10 +1,10 @@
 {
-  "name": "Glintless Potions",
-  "modId": "Glintless",
+  "name": "Modern Potions",
+  "modId": "ModernPotions",
   "author": "Syz66",
   "version": "2.0",
-  "link": "https://github.com/Syz66/GlintlessPotions",
+  "link": "https://github.com/Syz66/ModernPotions",
   "mixinConfigs": [
-    "glintless.mixins.json"
+    "modernpotions.mixins.json"
   ]
 }
